+++ packages/core/src/export.ts
@@ -48,6 +48,7 @@ export * from './behaviors/serviceAreaServer.js';
 export * from './behaviors/smokeCoAlarmServer.js';
 export * from './behaviors/switchServer.js';
 export * from './behaviors/thermostatServer.js';
+export * from './behaviors/timeSynchronizationServer.js';
 export * from './behaviors/valveConfigurationAndControlServer.js';
 export * from './behaviors/windowCoveringServer.js';
 export { addVirtualDevice } from './helpers.js';